+++ src/org/jivesoftware/spark/util/ResourceUtils.java
@@ -107,7 +107,7 @@ public final class ResourceUtils {
 
     public static String stripMnemonic(String label) {
         String text;
-        int index = label.indexOf("&");
+        int index = label.indexOf('&');
         if (index != -1) {
             text = label.substring(0, index);
             if (label.length() > index) {
@@ -119,9 +119,9 @@ public final class ResourceUtils {
     }
 
     public static int getMnemonicKeyCode(String mnemonic) {
-        int mindex = mnemonic.indexOf("&");
+        int mindex = mnemonic.indexOf('&');
         if (mindex > -1) {
-            return (int) mnemonic.toUpperCase().charAt(mindex + 1);
+            return mnemonic.toUpperCase().charAt(mindex + 1);
         }
         return 0;
     }
